+++ vue/components/molecules/thumbnails/thumbnails.vue
@@ -47,14 +47,14 @@ export const Thumbnails = {
         thumbnailFrame(thumbnail) {
             if (thumbnail.type === "video") return `video-${thumbnail.name}`;
             if (thumbnail.personalization) {
-                return `personalization-${thumbnail.face}-${thumbnail.frame}`;
+                return `personalization-[${thumbnail.group}]-${thumbnail.face}-${thumbnail.frame}`;
             }
             return `${thumbnail.face}-${thumbnail.frame}`;
         },
         thumbnailKey(thumbnail) {
             if (thumbnail.type === "video") return `video:${thumbnail.name}`;
             if (thumbnail.personalization) {
-                return `personalization-${thumbnail.frame}:${thumbnail.name}:${thumbnail.face}`;
+                return `personalization:${thumbnail.group}:${thumbnail.frame}:${thumbnail.name}:${thumbnail.face}`;
             }
             return `${thumbnail.frame}:${thumbnail.name}:${thumbnail.face}`;
         }